chore(gui): aligned location of mender specific components w/ nt-gui
- combined with https://github.com/NorthernTechHQ/nt-gui/pull/287/commits/82eec5d63642cae8f8a9cb24d2fd15379c61ff28
