+++ frontend/src/js/components/auditlogs/AuditLogs.tsx
@@ -17,8 +17,6 @@ import { useDispatch, useSelector } from 'react-redux';
 import { makeStyles } from 'tss-react/mui';
 
 import EnterpriseNotification, { DefaultUpgradeNotification } from '@northern.tech/common-ui/EnterpriseNotification';
-import { HELPTOOLTIPS } from '@northern.tech/common-ui/helptips/HelpTooltips';
-import { MenderHelpTooltip } from '@northern.tech/common-ui/helptips/MenderTooltip';
 import { AUDIT_LOGS_TYPES, BEGINNING_OF_TIME, BENEFITS, SORTING_OPTIONS, SP_AUDIT_LOGS_TYPES, TIMEOUTS } from '@northern.tech/store/constants';
 import { useLocationParams } from '@northern.tech/store/liststatehook';
 import {
@@ -36,6 +34,8 @@ import { createDownload, getISOStringBoundaries } from '@northern.tech/utils/hel
 import dayjs from 'dayjs';
 
 import historyImage from '../../../assets/img/history.png';
+import { HELPTOOLTIPS } from '../helptips/HelpTooltips';
+import { MenderHelpTooltip } from '../helptips/MenderTooltip';
 import AuditLogsFilter from './AuditLogsFilter';
 import AuditLogsList from './AuditLogsList';
 import AuditlogsView from './AuditlogsView';
