+++ frontend/src/js/components/deployments/deployment-wizard/ScheduleRollout.tsx
@@ -19,10 +19,11 @@ import { makeStyles } from 'tss-react/mui';
 
 import EnterpriseNotification from '@northern.tech/common-ui/EnterpriseNotification';
 import { InfoHintContainer } from '@northern.tech/common-ui/InfoHint';
-import { HELPTOOLTIPS } from '@northern.tech/common-ui/helptips/HelpTooltips';
-import { MenderHelpTooltip } from '@northern.tech/common-ui/helptips/MenderTooltip';
 import { BENEFITS } from '@northern.tech/store/constants';
 import dayjs from 'dayjs';
+
+import { HELPTOOLTIPS } from '../../helptips/HelpTooltips';
+import { MenderHelpTooltip } from '../../helptips/MenderTooltip';
 
 const useStyles = makeStyles()(() => ({
   textField: { minWidth: 400 },
