+++ frontend/src/js/components/devices/Groups.tsx
@@ -16,9 +16,10 @@ import { InfoOutlined as InfoIcon } from '@mui/icons-material';
 import { List, ListItemButton, ListItemIcon, ListItemText, ListSubheader } from '@mui/material';
 import { makeStyles } from 'tss-react/mui';
 
-import { HELPTOOLTIPS } from '@northern.tech/common-ui/helptips/HelpTooltips';
-import { MenderHelpTooltip } from '@northern.tech/common-ui/helptips/MenderTooltip';
 import { ALL_DEVICES } from '@northern.tech/store/constants';
+
+import { HELPTOOLTIPS } from '../helptips/HelpTooltips';
+import { MenderHelpTooltip } from '../helptips/MenderTooltip';
 
 const useStyles = makeStyles()(theme => ({
   header: {
