+++ frontend/src/js/components/devices/dialogs/DeviceConnectionDialog.test.tsx
@@ -15,8 +15,8 @@ import { act, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { vi } from 'vitest';
 
-import { defaultState, undefineds } from '../../../../tests/mockData';
-import { render } from '../../../../tests/setupTests';
+import { defaultState, undefineds } from '../../../../../tests/mockData';
+import { render } from '../../../../../tests/setupTests';
 import DeviceConnectionDialog from './DeviceConnectionDialog';
 
 describe('DeviceConnectionDialog Component', () => {
