+++ frontend/src/js/components/devices/dialogs/DeviceConnectionDialog.tsx
@@ -18,19 +18,19 @@ import { useNavigate } from 'react-router-dom';
 import { Button, DialogActions, DialogContent } from '@mui/material';
 import { makeStyles } from 'tss-react/mui';
 
+import DocsLink from '@northern.tech/common-ui/DocsLink';
+import InfoText from '@northern.tech/common-ui/InfoText';
+import Loader from '@northern.tech/common-ui/Loader';
+import { BaseDialog } from '@northern.tech/common-ui/dialogs/BaseDialog';
 import { DEVICE_STATES, TIMEOUTS, onboardingSteps } from '@northern.tech/store/constants';
 import { getDeviceCountsByStatus, getFeatures, getOnboardingState, getTenantCapabilities } from '@northern.tech/store/selectors';
 import { advanceOnboarding, saveUserSettings, setDeviceListState } from '@northern.tech/store/thunks';
 
-import docker from '../../../assets/img/docker.png';
-import raspberryPi4 from '../../../assets/img/raspberrypi4.png';
-import raspberryPi from '../../../assets/img/raspberrypi.png';
-import DocsLink from '../DocsLink';
-import InfoText from '../InfoText';
-import Loader from '../Loader';
-import { HELPTOOLTIPS } from '../helptips/HelpTooltips';
-import { MenderHelpTooltip } from '../helptips/MenderTooltip';
-import { BaseDialog } from './BaseDialog';
+import docker from '../../../../assets/img/docker.png';
+import raspberryPi4 from '../../../../assets/img/raspberrypi4.png';
+import raspberryPi from '../../../../assets/img/raspberrypi.png';
+import { HELPTOOLTIPS } from '../../helptips/HelpTooltips';
+import { MenderHelpTooltip } from '../../helptips/MenderTooltip';
 import PhysicalDeviceOnboarding from './PhysicalDeviceOnboarding';
 import VirtualDeviceOnboarding from './VirtualDeviceOnboarding';
