+++ frontend/src/js/components/devices/widgets/FilterItem.tsx
@@ -17,10 +17,10 @@ import { useCallback, useEffect, useRef, useState } from 'react';
 import { HighlightOff as HighlightOffIcon } from '@mui/icons-material';
 import { FormHelperText, IconButton, MenuItem, Select, TextField } from '@mui/material';
 
-import { HELPTOOLTIPS } from '@northern.tech/common-ui/helptips/HelpTooltips';
-import { MenderHelpTooltip } from '@northern.tech/common-ui/helptips/MenderTooltip';
 import { DEVICE_FILTERING_OPTIONS, TIMEOUTS, emptyFilter } from '@northern.tech/store/constants';
 
+import { HELPTOOLTIPS } from '../../helptips/HelpTooltips';
+import { MenderHelpTooltip } from '../../helptips/MenderTooltip';
 import AttributeAutoComplete from './AttributeAutocomplete';
 
 const filterOptionsByPlan = {
