+++ frontend/src/js/components/helptips/BaseOnboardingTip.test.tsx
