+++ frontend/src/js/components/helptips/BaseOnoardingTip.tsx
