+++ frontend/src/js/components/helptips/OnboardingCompleteTip.test.tsx
